requirements.txt
opencv-python==4.8.1.78
numpy==1.26.0
pyautogui==0.9.54
pytesseract==0.3.10
Pillow==10.0.1
mss==9.0.1
scikit-image==0.21.0
colorama==0.4.6
